+++ src/main/java/com/gamegoo/gamegoo_v2/common/validator/MemberValidator.java
@@ -1,7 +1,7 @@
 package com.gamegoo.gamegoo_v2.common.validator;
 
+import com.gamegoo.gamegoo_v2.exception.MemberException;
 import com.gamegoo.gamegoo_v2.exception.common.ErrorCode;
-import com.gamegoo.gamegoo_v2.exception.common.MemberException;
 import com.gamegoo.gamegoo_v2.member.domain.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
