--- src/main/java/com/gamegoo/gamegoo_v2/exception/common/MemberException.java
@@ -1,9 +1,0 @@
-package com.gamegoo.gamegoo_v2.exception.common;
-
-public class MemberException extends GlobalException {
-
-    public MemberException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-
-}
